Image Processing with TenserFlow
TensorFlow is a free and open source software library used for data flow and differentiable programming across a range of tasks. It is a symbolic math library and is also used for machine learning applications such as neural networks. The figure shows the object definition process with the Tenserflow library we used in our project.